Add Bacon.ajaxPoll
This provides a standard way to add polling of a json endpoint. Instead
of the existing behaviour of sending a request on an exact interval (ex.
one request every 300ms) this polls after an interval from the last
response.

This should fix issues when the server becomes overloaded and requests
get queued up.

Change-Id: I8a3eed7776e2334771766d53783d2348e66b8430