Switch to node-sass/libsass for CSS compilation
Ruby version of sass was burning tons of CPU for no real reason. This
helps.

node-sass is built on top of libsass, a C/C++ implementation of a Sass
compiler. I believe it might be still lacking a few features, but don't
think they'll cause issues here. It's also _much_ faster than the Ruby
version.

Also noticing that either with both version of Sass, Phoenix doesn't
kill the watchers after I stop the dev server. Not sure what's up with
that.

https://github.com/sass/libsass
https://github.com/sass/node-sass