Add waiting for start message to match warmup
Now that scoreboard instructions are always visible "pick service" seems
like an odd message for the warmup. I could have just sorted this out
with more CSS but it seemed that "waiting for match to start" was a
great message for this state anyways.

Also it opens the door to having important notifications at *any* time,
and not have them hidden by our styles for unrelated reasons.