Prevent scoreboard freakout when waiting
Because the ongoing match api 404s when there is no ongoing match, this
EMPTY_MATCH data is used to populate the scoreboard. Many of the things
on our scoreboard use $.fn.toggleClass to control an .active class. When
the property that controls them is simple undefined jQuery really does
toggle the class.