Set NODE_ENV in production webpack builds
This additional touch of configuration will replace instances of
`process.env.NODE_ENV` with `"production"`, when Webpack is run to
generate production builds.

This will turn any conditional code like the following:

    if (process.env.NODE_ENV === "production") {
      // do a thing
    } else {
      // do another thing
    }

Into:

    if ("production" === "production") {
      // do a thing
    } else {
      // do another thing
    }

During production builds, UglifyJS will take this one step further, see
that the conditional will always be `true`, and remove the exterior `if`
statement entirely, replacing it with the contents of the first block.