Move non-AR models out of app/models
While this could be a little controversial as many of these are not
really service objects, but I'd really prefer app/models contain only AR
models and potentially ActiveModel models. The question of what to do
with other domain objects and value objects remains, but I'm just
throwing them in app/services for now.